fifth_ed_spells/requirements.txt
Django==1.9.4

#Admin
Pillow==3.0.0
django-localflavor==1.2
django-import-export==0.4.2

#Static hosting
django-pipeline==1.6.0
libsass==0.10.0
cssmin==0.2.0
slimit==0.8.1
jsmin==2.2.0

#manifesto==0.5

#DB
psycopg2==2.6.1

#User Accounts
django-registration-redux==1.4
python-social-auth==0.2.14





#Utils
django-extensions==1.6.1
django-model-utils==2.4
django-braces==1.8.1
django-sslify==0.2.7
pytz==2015.7
easy-thumbnails==2.3

#JS tools
#django.js==0.8.1 # needs to be fixed up - using a version for now

django-casper==0.0.4

#REDIS CACHING
hiredis==0.2.0
redis==2.10.5
django-redis-cache==1.6.5
django-redis-sessions==0.5.0

## AWS
boto==2.38.0
django-storages==1.4
djangorestframework
django-filter
markdown

#HEROKU
dj-database-url==0.3.0
#dj-static==0.0.5
git+https://github.com/gone/dj-static.git@5da00529c033afa8e65aef13555b9eaec6dfc388#egg=dj-static
django-toolbelt==0.0.1
ipython==4.0.1
gunicorn==19.4.1

#OPTIONAL TOOLS
#whoosh
#django-haystack

#Money


stripe==1.12.0


#standard utilities
raven==5.9.2
#celery==3.1.18
sendgrid-django==1.4.0
